+++ Runtime/Utility/GameReady.cs
@@ -1,0 +1,17 @@
+﻿using System.Runtime.InteropServices;
+
+namespace Agava.YandexGames
+{
+    public static class GameReady
+    {
+        public static void NotifyLoadingCompleted()
+        {
+            GameReadyLoadingCompleted();
+        }
+
+        [DllImport("__Internal")]
+        private static extern void GameReadyLoadingCompleted();
+    }
+}
+
+