+++ Runtime/YandexGamesSdk.cs
@@ -9,6 +9,7 @@ namespace Agava.YandexGames
     public static class YandexGamesSdk
     {
         private static Action s_onInitializeSuccessCallback;
+        private static bool s_notifyLoadingCompletedOnInitialize;
 
         /// <summary>
         /// Enable it to log SDK callbacks in the console.
@@ -41,9 +42,10 @@ namespace Agava.YandexGames
         /// Downloads Yandex SDK script and inserts it into the HTML page.
         /// </summary>
         /// <returns>Coroutine waiting for <see cref="IsInitialized"/> to return true.</returns>
-        public static IEnumerator Initialize(Action onSuccessCallback = null)
+        public static IEnumerator Initialize(Action onSuccessCallback = null, bool callReadyOnCompletion = true)
         {
             s_onInitializeSuccessCallback = onSuccessCallback;
+            s_notifyLoadingCompletedOnInitialize = callReadyOnCompletion;
 
             YandexGamesSdkInitialize(OnInitializeSuccessCallback);
 
@@ -59,6 +61,9 @@ namespace Agava.YandexGames
         {
             if (CallbackLogging)
                 Debug.Log($"{nameof(YandexGamesSdk)}.{nameof(OnInitializeSuccessCallback)} invoked");
+
+            if (s_notifyLoadingCompletedOnInitialize)
+                GameReady.NotifyLoadingCompleted();
 
             s_onInitializeSuccessCallback?.Invoke();
         }